Fuzz iOS Test
Hello and welcome to the Fuzz Productions development test! The app you
will be developing should be a quick app to showcase skills that will
be valuable for the position you're applying for.

Upon launch, your app should query this link to provide content for the
app. The link is to a simple JSON file that contains the data for the
app. Format of the JSON should be self explanatory.

Create an app with a tabular navigation pattern. Containing the
following:

All - Display all content in a simple table/list view
Text - Display only text content in a table/list view
Images - Display only image content in a table/list view
Tapping any image item in the tables/list should open the image in a
full screen viewer. Tapping any text item in the tables/list should
navigate to a web view that has this site loaded.
That's it! There are no design requirements for this application;
simply use the built-in SDK components following design best practices.
Feel free to add bells and whistles.